Improves logic on GameBoard class and Player class
