+++ src/Logic/GameBoard.ts
@@ -1,4 +1,4 @@
-import { Player } from "./Player";
+import { Player } from "../Logic/Player";
 
 export class GameBoard {
   board: string[];
@@ -15,30 +15,32 @@ export class GameBoard {
     this.winner = "";
   }
 
-  isValidMove(boardPositionValue: string) {
+  isValidMove(boardPositionValue: string): boolean {
     return boardPositionValue === "";
   }
 
-  addMarker(player: Player, position: number) {
+  addMarker(
+    player: Player,
+    position: number
+  ): { status: boolean; message: string } {
     const validMove = this.isValidMove(this.board[position]);
 
     if (validMove) {
       this.marker = player.marker;
       this.board[position] = this.marker;
-      //TODO Add a function for the aiplayer to add a marker
+      return { status: validMove, message: "validMove" };
     } else {
-      console.log("invalid move");
-      return;
+      return { status: validMove, message: "invalidMove" };
     }
   }
 
-  resetBoard() {
-    this.board = Array(9).fill("");
+  resetBoard(): void {
+    this.board = Array(this.rows * this.columns).fill("");
     this.winner = "";
     this.marker = "";
   }
 
-  checkWin() {
+  checkGameStatus(): string | undefined {
     const winningCombinations = [
       [0, 1, 2],
       [3, 4, 5],
@@ -47,7 +49,7 @@ export class GameBoard {
       [1, 4, 7],
       [2, 5, 8],
       [0, 4, 8],
-      [2, 5, 6],
+      [2, 4, 6],
     ];
 
     for (const winningCombination of winningCombinations) {
@@ -56,12 +58,10 @@ export class GameBoard {
 
       if (board[a] !== "" && board[a] === board[b] && board[a] === board[c]) {
         this.winner = board[a];
-        return `winner is ${this.winner}`;
-      }
-      const isBoardFull = this.board.every((cell) => cell !== "");
-      if (isBoardFull) {
-        return "Draw";
+        return this.winner;
       }
     }
+    const isBoardFull = this.board.every((cell) => cell !== "");
+    return isBoardFull ? "Draw" : undefined;
   }
 }
